requirements.txt
# python modules
MySQLdb
dnspython
SubnetTree
psuntil >= 2.2